feat: add convenience link wrappers
