+++ package.json
@@ -39,6 +39,7 @@
   },
   "dependencies": {
     "h3": "^1.0.2",
+    "ofetch": "^1.0.0",
     "ohash": "^1.0.0",
     "ufo": "^1.0.0"
   },
